rewrite in C
C has a few advantages of shell for this.

First, it's much easier to robustly cleanup the lock file in various
failure conditions. I tried to get this behavior in shell, but it was
quite difficult.

Second, it doesn't force you to use any particular shell (the previous
implementation forced /bin/sh). Instead, it uses raw exec, and if you
want a shell, you have to pass your command through "SHELL -c". Although
this is somewhat regrettable, dealing with shell quoting issues seems
like a bad can of worms to open.

Third, it's just plain faster :).