shell exec functionality
This doesn't quite work, and I'm starting to think it's a bad idea.
Although execing a command through the shell lets you do nice things
like refer to env variables using $ and so on, it also means you have to
deal with lots of quoting issues because of turning an array of commands
into a single string to pass via $SHELL -c "SOME-COMMAND". This is very
nasty and likely shell-dependent, and it's probably a bad business to
get into.

Instead, if you really need to run the command through a shell, you can
always do:

solo $SHELL -c "COMMOND"

and deal with quoting issues yourself.